feat: improve spot swapper routing (#9327)
